% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asa.R
\name{asa}
\alias{asa}
\title{Model to solve the exact problem of Aircraft Selection and Allocation (ASA) for the containment of a wildfire.}
\usage{
asa(data, M_prime = 0, solver = "gurobi")
}
\arguments{
\item{data}{a list with the needed information about aircrafts and wildfire. For more information see \code{\link{example_data}} function.}

\item{M_prime}{penalization for the breach of the minimum aircrafts on a wildfire.}

\item{solver}{solver name, 'gurobi', 'Rsymphony' or 'lpSolve'.}
}
\value{
information about the selection and allocation of the aircrafts.
}
\description{
Model to solve the exact problem of Aircraft Selection and Allocation (ASA) for the containment of a wildfire.
}
\examples{
data <- asa::example_data()
asa::asa(data, solver="lpSolveAPI")
}
