% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{asa_dashboard}
\alias{asa_dashboard}
\title{Dashboard to run the Aircraft Selection and Allocation (ASA) package on the shiny app}
\usage{
asa_dashboard()
}
\value{
shinydashboard ui.
}
\description{
Dashboard to run the Aircraft Selection and Allocation (ASA) package on the shiny app
}
\examples{
SaA_dashboard()
}
