% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asa_h.R
\name{asa_h}
\alias{asa_h}
\title{Model to solve the problem of Aircraft Selection and Allocation (ASA) for the containment of a wildfire in a heuristic way.}
\usage{
asa_h(data, M_prime = 1e+06, niters = 10, solver = "gurobi")
}
\arguments{
\item{data}{a list with the needed information about aircrafts and wildfire. For more information see \code{\link{example_data}} function.}

\item{M_prime}{penalization for the breach of the minimum aircrafts on a wildfire.}

\item{niters}{maximum number of iterations.}

\item{solver}{solver name, 'gurobi', 'Rsymphony' or 'lpSolve'.}
}
\value{
information about the selection and allocation of the aircrafts.
}
\description{
Model to solve the problem of Aircraft Selection and Allocation (ASA) for the containment of a wildfire in a heuristic way.
}
\examples{
data <- example_data()
asa::asa_h(data, solver="lpSolveAPI")
}
