% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{asa_server}
\alias{asa_server}
\title{Server to run the Selection and Allocation (SaA) package on the shiny app}
\usage{
asa_server(input, output, session)
}
\value{
shiny server.
}
\description{
Server to run the Selection and Allocation (SaA) package on the shiny app
}
\examples{
asa_server
}
