% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-info_boxes.R
\name{box_init}
\alias{box_init}
\title{Info boxes initialization.}
\usage{
box_init(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
Info boxes initialization.
}
