% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_PER}
\alias{compute_PER}
\title{Compute the perimeter increment of the wildfire in each period.}
\usage{
compute_PER(SP, m)
}
\arguments{
\item{SP}{perimeter of the wildfire in each period.}

\item{m}{number of periods.}
}
\value{
perimeter increment.
}
\description{
Compute the perimeter increment of the wildfire in each period.
}
\examples{
data <-example_data()
compute_PER(data$SP, length(data$SP))
}
