% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contention_model.R
\name{contention_model}
\alias{contention_model}
\title{Contention model}
\usage{
contention_model(W_fix, S_fix, params, M, M_prime, solver = "gurobi")
}
\arguments{
\item{W_fix}{Work matrix.}

\item{S_fix}{Start matrix.}

\item{params}{model params.}

\item{M}{high value for the content_yes_no constraint.}

\item{M_prime}{penalty.}

\item{solver}{solver name. Options: 'gurobi', 'lpSolve' or 'Rsymphony'.}
}
\value{
optimal values.
}
\description{
Solve the contention model.
}
