% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{data.num.aircraft}
\alias{data.num.aircraft}
\title{Title}
\usage{
data.num.aircraft(sol)
}
\arguments{
\item{sol}{solution of asa problem.}
}
\value{
the number of aircraft in each period.
}
\description{
Title
}
