% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{data.scheduling.selection}
\alias{data.scheduling.selection}
\title{data.scheduling.selection}
\usage{
data.scheduling.selection(WRF, input)
}
\arguments{
\item{WRF}{matrix of 0's 1's 2's and 3's that represent the no-working, working, resting, and flying.}

\item{input}{shiny input.}
}
\value{
data
}
\description{
data.scheduling.selection
}
