% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_params}
\alias{get_model_params}
\title{get model paramas}
\usage{
get_model_params(data)
}
\arguments{
\item{data}{data with the aircraft, law and fire information}
}
\value{
SaA model params
}
\description{
get model paramas
}
\examples{
data <-example_data()
get_model_params(data)
}
