% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristic_model.R
\name{heuristic_model}
\alias{heuristic_model}
\title{Model to solve the problem of Aircraft Selection and Allocation (ASA) for the containment of a wildfire in a heuristic way.}
\usage{
heuristic_model(data, M_prime = 1e+06, niters = 10, solver = "gurobi",
  solver_params = list(TimeLimit = 600, OutputFlag = 0), verbose = 0)
}
\arguments{
\item{data}{a list with the needed information about aircrafts and wildfire. For more information see \code{\link{example_data}} function.}

\item{M_prime}{penalization for the breach of the minimum aircrafts on a wildfire.}

\item{niters}{maximum number of iterations.}

\item{solver}{solver name, 'gurobi', 'Rsymphony' or 'lpSolve'.}

\item{solver_params}{list of gurobi options. Defaults to list(TimeLimit=600, OutputFlag = 0).}

\item{verbose}{logging level information. If 0 no information.}
}
\value{
information about the selection and allocation of the aircrafts.
}
\description{
Model to solve the problem of Aircraft Selection and Allocation (ASA) for the containment of a wildfire in a heuristic way.
}
\examples{
data <- example_data()
asa::heuristic_model(data, solver="lpSolveAPI")
}
