% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infeasible_solution.R
\name{infeasible_solution}
\alias{infeasible_solution}
\title{Function that solves the rest model problem when the ASA model solution is infeasible.}
\usage{
infeasible_solution(data, solver, M_tilde = 1e+09,
  solver_params = list(TimeLimit = 600, OutputFlag = 0))
}
\arguments{
\item{data}{a list with the needed information about aircrafts and wildfire. For more information see \code{\link{example_data}} function.}

\item{solver}{solver name, 'gurobi', 'Rsymphony' or 'lpSolve'.}

\item{solver_params}{list of gurobi options. Defaults to list(TimeLimit=600, OutputFlag = 0).}

\item{M_prime}{penalization for the breach of the minimum aircrafts on a wildfire.}
}
\value{
information about the selection and allocation of the aircrafts.
}
\description{
Function that solves the rest model problem when the ASA model solution is infeasible.
}
\examples{
data <- asa::example_data()
asa::infeasible_solution(data, solver="lpSolveAPI")
}
