% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxes_initialization.R
\name{initialization}
\alias{initialization}
\title{Initialization.}
\usage{
initialization(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\description{
Initialization.
}
