% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_instances.R
\name{random_fire}
\alias{random_fire}
\title{Generate random set of wildfire periods.}
\usage{
random_fire(num_per, ini_SP_min = 1, ini_SP_max = 3, ini_NVC_min = 500,
  ini_NVC_max = 2000, rate_min = 1, rate_max = 1.3, EF_min = 0.5,
  EF_max = 1, nMin_min = numeric(num_per), nMin_max = rep(2, num_per))
}
\arguments{
\item{num_per}{number of periods.}

\item{ini_SP_min}{minimum value for the perimeter (SP) on the first period.}

\item{ini_SP_max}{maximum value for the perimeter (SP) on the first period.}

\item{ini_NVC_min}{minimum value for the wildfire cost (NVC) on the first period.}

\item{ini_NVC_max}{maximum value for the wildfire cost (NVC) on the first period.}

\item{rate_min}{minimum rate value to represent the evolution of the wildfire from a period to another.}

\item{rate_max}{maximum rate value to represent the evolution of the wildfire from a period to another.}

\item{EF_min}{minimum efficiency value of aircraft in each period.}

\item{EF_max}{maximum efficiency value of aircraft in each period.}

\item{nMin_min}{vector with the minimum number of minimum number of aircraft (nMin) in each period.}

\item{nMin_max}{vector with the maximum number of minimum number of aircraft (nMin) in each period.}
}
\value{
list of wildfire periods.
}
\description{
Generate random set of wildfire periods.
}
\examples{
random_fire(10)
}
