% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_instances.R
\name{random_instance}
\alias{random_instance}
\title{Generate random data.}
\usage{
random_instance(num_air, num_per)
}
\arguments{
\item{num_air}{number of aircrafts.}

\item{num_per}{number of periods.}
}
\value{
list with data information.
}
\description{
Generate random data.
}
\examples{
random_instance(5, 10)
}
