% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_instances.R
\name{random_law}
\alias{random_law}
\title{Generate random law.}
\usage{
random_law(nMax_min = 2, nMax_max = 8)
}
\arguments{
\item{nMax_min}{minimum number of maximum number of aircraft (nMax).}

\item{nMax_max}{maximum number of maximum number of aircraft (nMax).}
}
\value{
list with law information
}
\description{
Generate random law.
}
\examples{
random_law()
}
