% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest_model.R
\name{rest_model}
\alias{rest_model}
\title{Build and solve the rest model.}
\usage{
rest_model(I_select, m_select, S_fix_hist, params, M_prime, solver = "gurobi",
  solver_params = list(TimeLimit = 600, OutputFlag = 0))
}
\arguments{
\item{I_select}{selected aircraft.}

\item{m_select}{last period to estabish the aircraft rests.}

\item{S_fix_hist}{aircraft start period.}

\item{params}{common model params.}

\item{M_prime}{penalty for the aircraft number breach.}

\item{solver}{solver name: 'gurobi', 'lpSolve' or 'Rsymphony'.}

\item{solver_params}{list of gurobi options. Defaults to list(TimeLimit=600, OutputFlag = 0).}
}
\value{
list with the optimal value of the variables, objective function, ...
}
\description{
Mathematical model to solve the aircraft scheduling problem.
}
\examples{
data <- example_data()
params <- get_model_params(data)
I_select <- rep(T, length(params$I))
m_select <- length(params$Periods)
S_fix_hist = list()
M_prime = 1000000
rest_model(I_select, m_select, S_fix_hist, params, M_prime, solver="lpSolve")

}
