% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{shinyapp}
\alias{shinyapp}
\title{Shiny App for the Selection and Allocation problem.}
\usage{
shinyapp()
}
\value{
ShinyApp
}
\description{
Shiny App for the Selection and Allocation problem.
}
\examples{
shinyapp()
}
