% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{simulations}
\alias{simulations}
\title{Simulates and solves a battery of problems.}
\usage{
simulations(list_num_air, list_num_per, n_sims, seed = 1,
  outfile = "results.csv")
}
\arguments{
\item{list_num_air}{number of aircrafts list.}

\item{list_num_per}{number of periods list.}

\item{n_sims}{number of simulations.}

\item{seed}{seed number.}

\item{outfile}{output file.}
}
\value{
two csv files: ´´results.csv´´ with iterations information and ´´summary.csv´´ with a summary of these iterations results.
}
\description{
Simulates and solves a battery of problems.
}
\examples{
execute_instances(c(5, 10, 15), c(5, 10, 15), 10)
}
