% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{summary_table}
\alias{summary_table}
\title{Summarize simulation results.}
\usage{
summary_table(csv_file, out_file = "summary.csv")
}
\arguments{
\item{csv_file}{csv file input.}

\item{out_file}{csv file output}
}
\value{
csv with the summary of each instance.
}
\description{
Summarize simulation results.
}
