% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_nMin}
\alias{update_nMin}
\title{Update the minimum number of aircrafts in each period.}
\usage{
update_nMin(nMin, W_fix, FBRP, m)
}
\arguments{
\item{nMin}{Minimum number of aircrafts working in the wildfire in each time period.}

\item{W_fix}{work matrix.}

\item{FBRP}{Number of time periods flying from fire to rest place and vice versa.}

\item{m}{number of periods.}
}
\value{
updated nMin.
}
\description{
Update the minimum number of aircrafts in each period.
}
