% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{wrm_dashboard}
\alias{wrm_dashboard}
\title{Dashboard to run the Wildfire Resources Management (WRM) package on the shiny app}
\usage{
wrm_dashboard()
}
\value{
shinydashboard ui.
}
\description{
Dashboard to run the Wildfire Resources Management (WRM) package on the shiny app
}
\examples{
wrm_dashboard()
}
