% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asa.R
\name{asa}
\alias{asa}
\title{Model to solve the Aircraft Selection and Allocation problem (ASA) for the containment of a wildfire.}
\usage{
asa(data, M_prime = 0, method = "heuristic", niters = 10,
  solver = "gurobi", solver_params = list(TimeLimit = 600, OutputFlag = 0))
}
\arguments{
\item{data}{a list with the needed information about aircrafts and wildfire. For more information see \code{\link{example_data}} function.}

\item{M_prime}{penalization for the breach of the minimum aircrafts on a wildfire.}

\item{method}{desired method, exact or heuristic. Options 'exact' or 'heuristic'.}

\item{niters}{number of iterations for heuristic approach.}

\item{solver}{solver name, 'gurobi', 'Rsymphony' or 'lpSolve'.}

\item{solver_params}{list of gurobi options. Defaults to list(TimeLimit=600, OutputFlag = 0).}
}
\value{
information about the selection and allocation of the aircrafts.
}
\description{
Model to solve the Aircraft Selection and Allocation problem (ASA) for the containment of a wildfire.
}
\examples{
data <- asa::example_data()
asa::asa(data, solver="lpSolveAPI")
}
