% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server-solve.R
\name{solve_button}
\alias{solve_button}
\title{Solve button}
\usage{
solve_button(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}

\item{data}{}
}
\description{
Solve button
}
